% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build.qcr.object}
\alias{build.qcr.object}
\title{Define the object to hold QCR flags}
\usage{
build.qcr.object(rnaseq.data)
}
\arguments{
\item{rnaseq.data}{Seurat object}
}
\value{
QCR flag data.frame
}
\description{
Define the object to hold QCR flags
}
\keyword{internal}
